requirements.txt
chainlit
python-dotenv
-e .